sample_wsgidav.yaml: move settings for impersonator middleware 	to a separated section, not loading impersonator middleware 	by default wsgidav/impersonator.py: move user to uid/gid mapping logic here, 	update config entries, add more debug logging wsgidav/http_authenticator.py: remove impersonating-related logic
